tests/wpt/web-platform-tests/trusted-types/tt-block-eval.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <meta http-equiv="Content-Security-Policy" content="trusted-types *">
</head>
<body>
<script>
  trustedTypes.createPolicy("default", {createScript: _ => null});

  test(t => {
    let a = 0;
    assert_throws(new EvalError(), _ => {
      eval('a="hello there"');
    });
    assert_equals(a, 0);
  }, "eval blocks if the default policy rejects a value.");
</script>
